+++ src/main/java/cart/application/OrderService.java
@@ -17,11 +17,11 @@ import cart.domain.Member;
 import cart.domain.Order;
 import cart.domain.PointPolicy;
 import cart.domain.Product;
-import cart.dto.CartItemRequest;
-import cart.dto.CartPointsResponse;
-import cart.dto.OrderCreateRequest;
-import cart.dto.OrderItemResponse;
-import cart.dto.OrderResponse;
+import cart.dto.request.CartItemRequest;
+import cart.dto.response.CartPointsResponse;
+import cart.dto.request.OrderCreateRequest;
+import cart.dto.response.OrderItemResponse;
+import cart.dto.response.OrderResponse;
 import cart.entity.OrderEntity;
 import cart.entity.OrderItemEntity;
 import cart.entity.ProductEntity;
