+++ src/main/java/cart/application/ProductService.java
@@ -3,8 +3,8 @@ package cart.application;
 import cart.dao.CartItemDao;
 import cart.domain.Product;
 import cart.dao.ProductDao;
-import cart.dto.ProductRequest;
-import cart.dto.ProductResponse;
+import cart.dto.request.ProductRequest;
+import cart.dto.response.ProductResponse;
 import cart.entity.ProductEntity;
 
 import org.springframework.stereotype.Service;
