+++ src/main/java/cart/dto/request/CartItemAddRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.request;
 
 public class CartItemAddRequest {
     private Long productId;
