+++ src/main/java/cart/dto/response/CartItemUpdateResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.response;
 
 public class CartItemUpdateResponse {
     private final int quantity;
