+++ src/main/java/cart/dto/response/ProfileResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.response;
 
 public class ProfileResponse {
