+++ src/main/java/cart/ui/ProfileApiController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import cart.domain.Member;
-import cart.dto.ProfileResponse;
+import cart.dto.response.ProfileResponse;
 import io.swagger.v3.oas.annotations.Operation;
 
 @RestController
